[Bronze II] Title: 진법 변환, Time: 176 ms, Memory: 17752 KB -BaekjoonHub

+++ 백준/Bronze/2745. 진법 변환/진법 변환.java
@@ -7,6 +7,6 @@ public class Main {
         String N = sc.next();
         int B = sc.nextInt();
         
-        System.out.println(Integer.parseInt(N, B));
+        System.out.println(Long.parseLong(N, B));
     }
 }
